requirements.txt
yt-dlp
ffmpeg-python
dearpygui
Pillow
numpy